requirements.txt
# API
fastapi<0.110
uvicorn[standard]
SQLAlchemy
pydantic<2.0.0
python-multipart
requests
streamlit
streamlit-webrtc

# Database
pymysql
cryptography  # required for MySQL caching_sha2 auth

# CV/ML stack
numpy
scipy
opencv-python-headless
pillow
pyyaml
easydict
torch
torchvision
ultralytics
